багфикс для индексов массивов карты

+++ model.py
@@ -342,9 +342,9 @@ class Monsters(object):
                     dx = 0
                     dy = 0
                     if rnd(0,1):
-                        dy = (y - i)/abs(y - i)
+                        dy = int((y - i)/abs(y - i))
                     else:
-                        dx = (x - j)/abs(x - j)
+                        dx = int((x - j)/abs(x - j))
                     self.move(j, i, dx, dy)
                     obj.Canvas1.coords(self.monsters_map[i][j]['id'], (j+dx)*180+self.monsters_map[i][j]['coords'][0]*30+15, (i+dy)*180+self.monsters_map[i][j]['coords'][1]*30+30)
